requirements.txt
apache-airflow~=3.0.2
apache-airflow-providers-apache-kafka~=1.9.0